refactor for review
