+++ crates/core/src/table/state.rs
@@ -5,7 +5,6 @@ use std::sync::Arc;
 
 use arrow::compute::{filter_record_batch, is_not_null};
 use arrow_array::{Array, Int64Array, StringArray, StructArray};
-use arrow_cast::pretty::pretty_format_batches;
 use chrono::Utc;
 use futures::TryStreamExt;
 use object_store::{path::Path, ObjectStore};
@@ -16,7 +15,7 @@ use super::{get_partition_col_data_types, DeltaTableConfig};
 use crate::kernel::arrow::extract as ex;
 use crate::kernel::{
     Action, Add, DataType, EagerSnapshot, LogDataHandler, LogicalFile, Metadata, Protocol, Remove,
-    ReplayVistor, StructType,
+    ReplayVisitor, StructType,
 };
 use crate::logstore::LogStore;
 use crate::operations::transaction::CommitData;
@@ -47,7 +46,7 @@ impl AppTransactionVisitor {
     }
 }
 
-impl ReplayVistor for AppTransactionVisitor {
+impl ReplayVisitor for AppTransactionVisitor {
     fn visit_batch(&mut self, batch: &arrow_array::RecordBatch) -> DeltaResult<()> {
         if batch.column_by_name("txn").is_none() {
             return Ok(());
@@ -90,7 +89,7 @@ impl DeltaTableState {
         version: Option<i64>,
     ) -> DeltaResult<Self> {
         let mut app_visitor = AppTransactionVisitor::new();
-        let visitors: Vec<&mut dyn ReplayVistor> = vec![&mut app_visitor];
+        let visitors: Vec<&mut dyn ReplayVisitor> = vec![&mut app_visitor];
         let snapshot = EagerSnapshot::try_new_with_visitor(
             table_root,
             store.clone(),
@@ -284,7 +283,6 @@ impl DeltaTableState {
         log_store: Arc<dyn LogStore>,
         version: Option<i64>,
     ) -> Result<(), DeltaTableError> {
-        println!("update table");
         let mut app_txn_visitor = AppTransactionVisitor::new();
         self.snapshot
             .update(log_store, version, vec![&mut app_txn_visitor])
